Rename `cbio` to `cbioportal`
