+++ csr2cbioportal/csr2cbioportal.py
@@ -13,9 +13,9 @@ import click
 from csr.csr import CentralSubjectRegistry
 from csr.study_registry_reader import SubjectRegistryReader
 
-from .cbio_transform_clinical import write_clinical, transform_patient_clinical_data, transform_sample_clinical_data
-from .cbio_create_metafile import create_meta_content
-from .cbio_create_caselist import create_caselist
+from .transform_clinical import write_clinical, transform_patient_clinical_data, transform_sample_clinical_data
+from .create_metafile import create_meta_content
+from .create_caselist import create_caselist
 import pandas as pd
 import shutil
 import time
@@ -39,7 +39,7 @@ DESCRIPTION = 'Transformed to cBioPortal format on: %s' % time.strftime("%d-%m-%
 TYPE_OF_CANCER = 'mixed'
 
 
-def create_cbio_study(input_dir, ngs_dir, output_dir):
+def create_cbioportal_study(input_dir, ngs_dir, output_dir):
     # Remove old output directory and recreate to ensure all NGS data is
     if os.path.exists(output_dir):
         shutil.rmtree(output_dir)
@@ -303,7 +303,7 @@ def get_complete_header(paths_to_process):
 @click.argument('output_dir')
 @click.version_option()
 def run(input_dir, ngs_dir, output_dir):
-    create_cbio_study(input_dir, ngs_dir, output_dir)
+    create_cbioportal_study(input_dir, ngs_dir, output_dir)
 
 
 def main():
