+++ setup.cfg
@@ -7,7 +7,7 @@ test=pytest
 
 [coverage:run]
 branch = True
-source = csr2transmart, csr2cbio, sources2csr, csr
+source = csr2transmart, csr2cbioportal, sources2csr, csr
 
 [tool:pytest]
 testpaths = tests
