+++ setup.py
@@ -35,13 +35,13 @@ setup(
         'csr2transmart',
         'csr2transmart.mappers',
         'sources2csr',
-        'csr2cbio',
+        'csr2cbioportal',
         'csr'
     ],
     entry_points={
         'console_scripts': ['csr2transmart=csr2transmart.csr2transmart:main',
                             'sources2csr=sources2csr.sources2csr:main',
-                            'csr2cbio=csr2cbio.csr2cbio:main'],
+                            'csr2cbioportal=csr2cbioportal.csr2cbioportal:main'],
     },
     include_package_data=True,
     license="MIT",
